Provisioning: Share one Grafana server per package in jobs integration tests
Apply the same shared-server strategy from the repository package to
the jobs package. One Grafana server is started via sync.Once and reused
by all TestIntegration* functions, with CleanupAllResources running
between tests for isolation.

Tests that require incompatible server configurations are moved to
dedicated sub-packages so they get their own database and server:

- jobs/disabledfeatures/ — export feature flag disabled
- jobs/conflict/ — controllers disabled (tests optimistic locking)
- jobs/flagdisabled/ — folder metadata flag disabled

This reduces server startups from ~30 to 4 (1 shared + 3 isolated).

Made-with: Cursor
