+++ pkg/tests/apis/provisioning/jobs/conflict/conflict_test.go
@@ -1,4 +1,4 @@
-package jobs
+package conflict
 
 import (
 	"context"
@@ -10,23 +10,14 @@ import (
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 
 	provisioning "github.com/grafana/grafana/apps/provisioning/pkg/apis/provisioning/v0alpha1"
-	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
-	"github.com/grafana/grafana/pkg/tests/testinfra"
-	"github.com/grafana/grafana/pkg/util/testutil"
 )
 
 // TestIntegrationProvisioning_JobConflict tests that if two concurrent drivers try to update a
-// job they received before the other updated it, that one will fail. This is critical for concurrent jobs
+// job they received before the other updated it, that one will fail. This is critical for concurrent jobs.
 func TestIntegrationProvisioning_JobConflict(t *testing.T) {
-	testutil.SkipIntegrationTestInShortMode(t)
-
-	// disable the controllers so the jobs don't get auto-processed
-	helper := common.RunGrafana(t, func(opts *testinfra.GrafanaOpts) {
-		opts.DisableControllers = true
-	})
+	helper := sharedHelper(t)
 	ctx := context.Background()
 
-	// create a job
 	obj := &unstructured.Unstructured{
 		Object: map[string]interface{}{
 			"apiVersion": "provisioning.grafana.app/v0alpha1",
@@ -47,14 +38,11 @@ func TestIntegrationProvisioning_JobConflict(t *testing.T) {
 	createdJob, err := helper.Jobs.Resource.Create(ctx, obj, metav1.CreateOptions{})
 	require.NoError(t, err)
 
-	// have two clients get the same job before either has updated it. this simulates the race condition
-	// between two concurrent workers.
 	job, err := helper.Jobs.Resource.Get(ctx, createdJob.GetName(), metav1.GetOptions{})
 	require.NoError(t, err)
 	job2, err := helper.Jobs.Resource.Get(ctx, createdJob.GetName(), metav1.GetOptions{})
 	require.NoError(t, err)
 
-	// have the first client update the job, this should update the RV
 	client1Update := job.DeepCopy()
 	if client1Update.GetLabels() == nil {
 		client1Update.SetLabels(make(map[string]string))
@@ -66,7 +54,6 @@ func TestIntegrationProvisioning_JobConflict(t *testing.T) {
 	require.NoError(t, err)
 	require.NotEqual(t, job.GetResourceVersion(), updatedJob.GetResourceVersion())
 
-	// now when client two tries to update the job, the RV is no longer what it originally received
 	client2Update := job2.DeepCopy()
 	if client2Update.GetLabels() == nil {
 		client2Update.SetLabels(make(map[string]string))
@@ -78,7 +65,6 @@ func TestIntegrationProvisioning_JobConflict(t *testing.T) {
 	require.Error(t, err)
 	require.True(t, apierrors.IsConflict(err), "should get conflict error when updating with stale resource version")
 
-	// clean up
 	err = helper.Jobs.Resource.Delete(ctx, createdJob.GetName(), metav1.DeleteOptions{})
 	require.NoError(t, err)
 }
