+++ pkg/tests/apis/provisioning/jobs/conflict/helpers_test.go
@@ -1,0 +1,21 @@
+package conflict
+
+import (
+	"testing"
+
+	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
+	"github.com/grafana/grafana/pkg/tests/testinfra"
+)
+
+var env = common.NewSharedEnv(func(opts *testinfra.GrafanaOpts) {
+	opts.DisableControllers = true
+})
+
+func sharedHelper(t *testing.T) *common.ProvisioningTestHelper {
+	t.Helper()
+	return env.GetCleanHelper(t)
+}
+
+func TestMain(m *testing.M) {
+	env.RunTestMain(m)
+}
