+++ pkg/tests/apis/provisioning/jobs/flagdisabled/helpers_test.go
@@ -1,0 +1,18 @@
+package flagdisabled
+
+import (
+	"testing"
+
+	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
+)
+
+var env = common.NewSharedEnv()
+
+func sharedHelper(t *testing.T) *common.ProvisioningTestHelper {
+	t.Helper()
+	return env.GetCleanHelper(t)
+}
+
+func TestMain(m *testing.M) {
+	env.RunTestMain(m)
+}
