+++ pkg/tests/apis/provisioning/jobs/job_warning_result_test.go
@@ -10,13 +10,10 @@ import (
 
 	provisioning "github.com/grafana/grafana/apps/provisioning/pkg/apis/provisioning/v0alpha1"
 	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
-	"github.com/grafana/grafana/pkg/util/testutil"
 )
 
 func TestIntegrationProvisioning_JobWarningResult(t *testing.T) {
-	testutil.SkipIntegrationTestInShortMode(t)
-
-	helper := common.RunGrafana(t)
+	helper := sharedHelper(t)
 
 	// Create a test repository with a malformed dashboard file
 	const repo = "job-warning-test-repo"
@@ -71,9 +68,7 @@ func TestIntegrationProvisioning_JobWarningResult(t *testing.T) {
 }
 
 func TestIntegrationProvisioning_JobWarningResult_MissingName(t *testing.T) {
-	testutil.SkipIntegrationTestInShortMode(t)
-
-	helper := common.RunGrafana(t)
+	helper := sharedHelper(t)
 
 	// Create a test repository with a dashboard file missing the name field
 	const repo = "job-warning-missing-name-repo"
@@ -119,9 +114,7 @@ func TestIntegrationProvisioning_JobWarningResult_MissingName(t *testing.T) {
 }
 
 func TestIntegrationProvisioning_JobWarningResult_DashboardRefreshInterval(t *testing.T) {
-	testutil.SkipIntegrationTestInShortMode(t)
-
-	helper := common.RunGrafana(t)
+	helper := sharedHelper(t)
 
 	// Create a test repository with a dashboard file with refresh interval too low
 	const repo = "job-warning-refresh-interval-repo"
@@ -167,9 +160,7 @@ func TestIntegrationProvisioning_JobWarningResult_DashboardRefreshInterval(t *te
 }
 
 func TestIntegrationProvisioning_JobWarningResult_DuplicateName(t *testing.T) {
-	testutil.SkipIntegrationTestInShortMode(t)
-
-	helper := common.RunGrafana(t)
+	helper := sharedHelper(t)
 
 	// Create a test repository with two dashboard files that share the same metadata.name.
 	// The second file processed should trigger a "duplicate resource name" validation warning.
