+++ pkg/tests/apis/provisioning/jobs/migratejob_test.go
@@ -17,57 +17,13 @@ import (
 	provisioning "github.com/grafana/grafana/apps/provisioning/pkg/apis/provisioning/v0alpha1"
 	"github.com/grafana/grafana/pkg/apimachinery/utils"
 	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
-	"github.com/grafana/grafana/pkg/util/testutil"
 )
-
-func TestIntegrationProvisioning_MigrateDisabledByConfiguration(t *testing.T) {
-	testutil.SkipIntegrationTestInShortMode(t)
-
-	// Run Grafana WITHOUT the export feature flag enabled
-	helper := common.RunGrafana(t, common.WithoutExportFeatureFlag)
-
-	// Create a repository
-	const repo = "test-repository"
-	testRepo := common.TestRepo{
-		Name:   repo,
-		Target: "instance",
-	}
-	helper.CreateRepo(t, testRepo)
-
-	// Try to trigger a migrate job (it should fail)
-	spec := provisioning.JobSpec{
-		Action: provisioning.JobActionMigrate,
-		Migrate: &provisioning.MigrateJobOptions{
-			Message: "Test migration",
-		},
-	}
-
-	// Trigger job and wait for it to complete (will fail)
-	job := helper.TriggerJobAndWaitForComplete(t, repo, spec)
-
-	// Check that job failed with the expected error message
-	status, found, err := unstructured.NestedMap(job.Object, "status")
-	require.NoError(t, err)
-	require.True(t, found, "job should have status")
-
-	state, found, err := unstructured.NestedString(status, "state")
-	require.NoError(t, err)
-	require.True(t, found)
-	require.Equal(t, "error", state, "job should have error state")
-
-	message, found, err := unstructured.NestedString(status, "message")
-	require.NoError(t, err)
-	require.True(t, found)
-	require.Contains(t, message, "migrate functionality is disabled by configuration")
-}
 
 // TestIntegrationProvisioning_PullJobUnmanagedConflict verifies that a pull/sync
 // job produces a warning when it encounters a pre-existing unmanaged resource
 // with the same name as a file in the repository.
 func TestIntegrationProvisioning_PullJobUnmanagedConflict(t *testing.T) {
-	testutil.SkipIntegrationTestInShortMode(t)
-
-	helper := common.RunGrafana(t)
+	helper := sharedHelper(t)
 	ctx := context.Background()
 
 	// Step 1: Create an unmanaged dashboard directly via the API.
@@ -151,9 +107,7 @@ func TestIntegrationProvisioning_PullJobUnmanagedConflict(t *testing.T) {
 // 2. The Sync phase takes them over because they appear in the takeover allowlist.
 // 3. After migration, the dashboards are managed by the repository.
 func TestIntegrationProvisioning_MigrateTakeover(t *testing.T) {
-	testutil.SkipIntegrationTestInShortMode(t)
-
-	helper := common.RunGrafana(t)
+	helper := sharedHelper(t)
 	ctx := context.Background()
 
 	// Step 1: Create two unmanaged dashboards directly.
@@ -239,9 +193,7 @@ func TestIntegrationProvisioning_MigrateTakeover(t *testing.T) {
 // Dashboards already claimed by the first repository are skipped.
 // This test focuses on files written to each repository directory.
 func TestIntegrationProvisioning_SecondMigrateOnlyExportsNewDashboards(t *testing.T) {
-	testutil.SkipIntegrationTestInShortMode(t)
-
-	helper := common.RunGrafana(t)
+	helper := sharedHelper(t)
 	ctx := context.Background()
 
 	// Create two unmanaged dashboards.
