+++ pkg/tests/apis/provisioning/jobs/orphanjob_get_test.go
@@ -12,7 +12,6 @@ import (
 
 	provisioning "github.com/grafana/grafana/apps/provisioning/pkg/apis/provisioning/v0alpha1"
 	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
-	"github.com/grafana/grafana/pkg/util/testutil"
 )
 
 // TestIntegrationProvisioning_OrphanCleanupJobGetEndpoint verifies that the
@@ -20,9 +19,7 @@ import (
 // when the repository does not exist. This is critical because these jobs are
 // specifically created for missing or terminating repositories.
 func TestIntegrationProvisioning_OrphanCleanupJobGetEndpoint(t *testing.T) {
-	testutil.SkipIntegrationTestInShortMode(t)
-
-	helper := common.RunGrafana(t)
+	helper := sharedHelper(t)
 	ctx := context.Background()
 
 	for _, action := range []provisioning.JobAction{
